refactor: Move ggshield.core.file_utils to ggshield.utils.files
The functions inside the module are not ggshield specific and should be
in the ggshield.utils module according to the architecture described in
 #521.

refs #521
