+++ ggshield/cmd/iac/scan/diff.py
@@ -23,10 +23,10 @@ from ggshield.cmd.utils.common_options import (
 )
 from ggshield.core.config import Config
 from ggshield.core.errors import handle_exception
-from ggshield.core.file_utils import is_filepath_excluded
 from ggshield.core.scan import ScanContext, ScanMode
 from ggshield.core.tar_utils import INDEX_REF, get_empty_tar
 from ggshield.core.text_utils import display_info, display_warning
+from ggshield.utils.files import is_filepath_excluded
 from ggshield.utils.git_shell import (
     Filemode,
     get_diff_files_status,
